backend/requirements.txt
Django==5.0.10
djangorestframework==3.15.2
gunicorn==22.0.0
psycopg[binary]==3.1.19
djangorestframework-simplejwt==5.3.1
djangorestframework-simplejwt[crypto]==5.3.1